test: move termynal directive-helper tests out of test_plugin.py
The _as_bool/_as_int coverage tests landed in test_plugin.py (a general
plugin test file); relocate them to a termynal-scoped test_termynal_directive.py
so the termynal feature keeps its own test footprint. test_plugin.py is
restored to its prior state; no behavior change.
